+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 
 ## Upcoming Release
+* Added cmdlets for group owner
+    - `Get-AzADGroupOwner`
+    - `New-AzADGroupOwner`
+    - `Remove-AzADGroupOwner`
 
 ## Version 6.8.0
 * Fixed the incorrect behavior of pagination for `Get-AzTag`