+++ src/Resources/Resources/ChangeLog.md
@@ -19,10 +19,12 @@
 -->
 
 ## Upcoming Release
+* Fixed the issue that `New-AzRoleAssignment` didn't work without subscription.
 * Added cmdlets for group owner
     - `Get-AzADGroupOwner`
     - `New-AzADGroupOwner`
     - `Remove-AzADGroupOwner`
+
 
 ## Version 6.8.0
 * Fixed the incorrect behavior of pagination for `Get-AzTag`