+++ src/Resources/Resources/ChangeLog.md
@@ -24,7 +24,7 @@
     - `Get-AzADGroupOwner`
     - `New-AzADGroupOwner`
     - `Remove-AzADGroupOwner`
-
+* Updated Tags functionality in deployment stacks New and Set cmdlets
 
 ## Version 6.8.0
 * Fixed the incorrect behavior of pagination for `Get-AzTag`